fix firefox link opening
Running from xprofile appears to bypass the dbus session that xfce is
started with (aka no DBUS_ env vars), and this is then inherited by
any commands launched by xbindkeys.  This is a problem as new-er
firefox exclusively uses DBUS for merging sessions together and if it
can't find the bus you will be crossing between "xbindkeys procs" and
"everything else" results in the dreaded "firefox is already running"
error.